Small refactoring.
Moved `Dataset` and `Joke` into `data.py`.
