--- bot/joke.py
@@ -1,4 +1,0 @@
-class Joke:
-    def __init__(self, text, id):
-        self.id = id
-        self.text = text
